Reverted back to not using a separate object for initializing a service
