+++ app/src/main/java/com/michaelcorral/ortimer/screens/mainscreen/MainScreenPresenter.kt
@@ -21,7 +21,7 @@ class MainScreenPresenter(
         // TODO: Refactor to factory or some kind of polymorphism
         toggleSession = repository.retrieveSessionState()
         if (toggleSession) {
-            stopSession()
+//            stopSession()
             startSession()
         }
